requirements.txt
Django
yt-dlp
moviepy
google-generativeai
webvtt-py
google-api-python-client
google-auth-oauthlib
gunicorn